% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_IRAM.R
\name{Arnoldi}
\alias{Arnoldi}
\title{A implementation of Arnoldi Decomposition}
\usage{
Arnoldi(A, b, V0, H0, r0, n, m, k, First = FALSE)
}
\arguments{
\item{A}{input matrix, to decompose it into AV = VH + r*t(e)}

\item{b}{vector for Krylov-subspace}

\item{V0}{start V}

\item{H0}{start H}

\item{r0}{start r}

\item{n}{dimension of A}

\item{m}{dimension of H}

\item{k}{need k largest eigenvalues}

\item{First}{whether it is the first Arnoldi, before any restart}
}
\value{
A list containing H, V and r
}
\description{
A implementation of Arnoldi Decomposition
}
