% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_KSLE.R
\name{Arnoldi_Krylov}
\alias{Arnoldi_Krylov}
\title{A implementation of reducing the the Krylov form S to Hessenberg form by Householder matrix}
\usage{
Arnoldi_Krylov(U, S, u, b, m, k, tol0 = 1e-08)
}
\arguments{
\item{U}{n*m orthogonal matrix, U of AU = US + ub}

\item{S}{m*m matrix, S of AU = US + ub}

\item{u}{n*1 matrix, u of AU = US + ub}

\item{b}{1*m matrix, b of AU = US + ub}

\item{m}{dimension of S}

\item{k}{need k largest eigenvalues}

\item{tol0}{tolerance for setting small numbers to 0}
}
\value{
A list containing U, S, u and b, where S is tridiagonal matrix
}
\description{
A implementation of reducing the the Krylov form S to Hessenberg form by Householder matrix
}
