% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_KSLE.R
\name{check_converge}
\alias{check_converge}
\title{A implementation of checking convergence of Krylov-Schur Decomposition}
\usage{
check_converge(A, V, H, G, n, k, tol = 1e-06)
}
\arguments{
\item{A}{input matrix}

\item{V}{the orthogonal matrix of AV = VS + ub}

\item{H}{the diagonal matrix of Schur decomposition of S}

\item{G}{the orthogonal matrix of Schur decomposition of S}

\item{n}{dimension of A}

\item{k}{need k largest eigenvalues}

\item{tol}{tolerance of convergence}
}
\value{
A list containing convergence and eigenpairs
}
\description{
A implementation of checking convergence of Krylov-Schur Decomposition
}
