% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_IRAM.R
\name{eigen_IRAM}
\alias{eigen_IRAM}
\title{A implementation IRAM for EVD}
\usage{
eigen_IRAM(A, k, max_iter = 10000)
}
\arguments{
\item{A}{input matrix}

\item{k}{need k largest eigenvalues}

\item{max_iter}{maximum of iteration number}
}
\value{
A list containing eigenvalues and eigenvectors
}
\description{
A implementation IRAM for EVD
}
