% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_KSLE.R
\name{eigen_KSLE}
\alias{eigen_KSLE}
\title{A implementation Krylov-Schur Algorithm for EVD}
\usage{
eigen_KSLE(A, k, max_iter = 10000)
}
\arguments{
\item{A}{input matrix}

\item{k}{need k largest eigenvalues}

\item{max_iter}{maximum of iteration number}
}
\value{
A list containing eigenvalues and eigenvectors
}
\description{
A implementation Krylov-Schur Algorithm for EVD
}
