% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen_IRAM.R
\name{ImplicitlyRestart}
\alias{ImplicitlyRestart}
\title{A implementation of Implicitly Restart Arnoldi}
\usage{
ImplicitlyRestart(H, V, r, eigenVal, k, m)
}
\arguments{
\item{H}{Hessenberg matrix, H of AV = VH + r*t(e)}

\item{V}{orthogonal matrix, V of AV = VH + r*t(e)}

\item{r}{vector, r of AV = VH + r*t(e)}

\item{eigenVal}{eigenvalues of H}

\item{k}{need k largest eigenvalues}

\item{m}{dimension of H}
}
\value{
A list containing H, V and r after restarting
}
\description{
A implementation of Implicitly Restart Arnoldi
}
