% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompositions_utility.R
\name{SchurDecom}
\alias{SchurDecom}
\title{Call Schur() to decompose and sort}
\usage{
SchurDecom(H)
}
\arguments{
\item{H}{input matrix}
}
\value{
A list containing Q and T for H = QTQ'
}
\description{
Call Schur() to decompose and sort
}
\examples{
H <- matrix(c(1,2,3,4,2,5,6,7,3,6,8,9,4,7,9,10), 4, 4)
SchurDecom(H)
}
