% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncatedSVD.R
\name{truncatedSVD}
\alias{truncatedSVD}
\title{A implementation of SVD, both IRAM and Krylov-Schur algorithm}
\usage{
truncatedSVD(M, k, method = "KSLE", Uonly = FALSE)
}
\arguments{
\item{M}{input matrix}

\item{k}{need k singular values}

\item{method}{choice includes "IRAM" and "KSLE"}

\item{Uonly}{whether only need U and D}
}
\value{
A list containing U, D and V
}
\description{
A implementation of SVD, both IRAM and Krylov-Schur algorithm
}
